Update to be more in line with VSCode
Updated the mappings to be more in line with VSCode, additionally changed some of the default vim behavior to ensure that the macros in the editor also behave the same was as in VSCode.

Updated the telescope to always use the root of the repository as its root to search through files. This will likely have to be updated at a later time to ensure that it always uses the root of the session instead of the root of the git repository.